// NSP/src/app/Models/Colaboradores/colaborador.model.ts
export interface Colaboradores {
    id?: String;         
    nome: String;
    email: String;
    nascimento: Date
    funcao: String;
    logradouro: String;
    numero: String;
    complemento: String;
    cep: String;
    bairro: String;
    cidade: String;
    telefone: String;
    celular: String;
    uf: String;
}